ultimate_project/api_gateway/requirements.txt
fastapi
uvicorn
httpx
websockets
django-cors-headers
cryptography
PyJWT
requests
python-multipart
